+++ src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -22,7 +22,7 @@ public class FileBasedDatabase implements Database {
   private final JsonService jsonService;
 
   @Override
-  public int save(Invoice invoice) {
+  public long save(Invoice invoice) {
     try {
       invoice.setId(idProvider.getNextIdAndIncrement());
       filesService.appendLineToFile(databasePath, jsonService.toJson(invoice));
@@ -34,7 +34,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> getById(int id) {
+  public Optional<Invoice> getById(long id) {
     try {
       return filesService.readAllLines(databasePath)
           .stream()
@@ -47,7 +47,7 @@ public class FileBasedDatabase implements Database {
     }
   }
 
-  private boolean containsId(String line, int id) {
+  private boolean containsId(String line, long id) {
     return line.contains("{\"id\":" + id + ",\"number\"");
   }
 
@@ -65,7 +65,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+  public Optional<Invoice> update(long id, Invoice updatedInvoice) {
     try {
       List<String> allInvoices = filesService.readAllLines(databasePath);
       var invoicesWithoutInvoiceWithGivenId = allInvoices
@@ -84,7 +84,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> delete(int id) {
+  public Optional<Invoice> delete(long id) {
     try {
       var allInvoices = filesService.readAllLines(databasePath);
       var invoicesExceptDeleted = allInvoices
